requirements.txt
aioflask==0.4.0
asgiref==3.7.2
astroid==2.15.5
asyncio==3.4.3
black==23.3.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.7
cryptography==41.0.1
dill==0.3.6
exceptiongroup==1.1.2
Flask==2.3.2
greenlet==2.0.2
greenletio==0.10.1
h11==0.14.0
idna==3.4
importlib-metadata==6.7.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
pluggy==1.2.0
pycparser==2.21
PyJWT==2.7.0
pylint==2.17.4
pytest==7.4.0
pytest-cov==4.1.0
requests==2.31.0
tomli==2.0.1
tomlkit==0.11.8
typing==3.7.4.3
typing-extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.15.0
